% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dow.R
\docType{data}
\name{dow}
\alias{dow}
\title{DOWSING data set}
\format{
A data frame with 26 observations and 4 columns:
\describe{
  \item{TRIAL}{Trial number}
  \item{DOWSER}{Assigned number to individual dowsers}
  \item{PIPE}{Actual pipe location}
  \item{GUESS}{Dowsers guess}
}
}
\source{
<https://www.amazon.com/Statistics-Engineering-Sciences-William...
}
\usage{
dow
}
\description{
A data set of data on whether dowsers can actually divine water
}
\keyword{datasets}
