% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knotty_rsq.R
\name{knotty_rsq}
\alias{knotty_rsq}
\title{knotty_rsq}
\usage{
knotty_rsq(df, nval = 1000)
}
\arguments{
\item{df}{data frame}

\item{nval}{used for the length of the `seq()` and `vector()` functions}
}
\value{
a knot value `xk` which maximizes `Rsquared`
}
\description{
Calculates the knot value `xk` which maximizes `Rsquared`
}
\examples{
\dontrun{knotty_rsq(df,nval=1000)}
}
