% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myknot.R
\name{myknot}
\alias{myknot}
\title{myknot}
\usage{
myknot(xk, x, y)
}
\arguments{
\item{xk}{chosen knot value}

\item{x}{explanatory variable}

\item{y}{response variable}
}
\value{
`Rsquared` value
}
\description{
Calculates the `Rsquared` value for a given `xk`
}
\examples{
\dontrun{myknot(18,x,y)}
}
