% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogplot.R
\name{ogplot}
\alias{ogplot}
\title{ogplot}
\usage{
ogplot(xk)
}
\arguments{
\item{xk}{knot value}
}
\value{
a plot of the data with fitted piecewise regression model
}
\description{
The scatterplot of the data with fitted piecewise regression model
}
\examples{
\dontrun{ogplot(xk = 18)}
}
