% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyspruce.R
\name{shinyspruce}
\alias{shinyspruce}
\title{shinyspruce}
\usage{
shinyspruce()
}
\value{
two plots
}
\description{
This Shiny App investigates a piecewise regression analyses on plots based on the Spruce data set.
}
\examples{
\dontrun{shinyspruce()}
}
