% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spruce.df.R
\docType{data}
\name{spruce.df}
\alias{spruce.df}
\title{SPRUCE data set}
\format{
A data frame with 36 rows and 2 vector columns:
\describe{ **EDIT FIELDS BELOW FOR VARIABLES**
  \item{BHDiameter}{Breast-Height Diameter of the tree}
  \item{Height}{The actual height of the tree}
  ...
}
}
\source{
<https://www.amazon.com/Statistics-Engineering-Sciences-William...
}
\usage{
spruce.df
}
\description{
A data of Spruce trees.
}
\keyword{datasets}
