% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myknot.R
\name{myknot}
\alias{myknot}
\title{myknot}
\usage{
myknot(x_k, x, y, data)
}
\arguments{
\item{x_k}{knot value}

\item{x}{explanatory variable}

\item{y}{response variable}

\item{data}{data frame}
}
\value{
`r.squared` from the give knot value `x_k`
}
\description{
Calculates `r.squared` from the give knot value `x_k`
}
\examples{
\dontrun{myknot(17, x=x, y=y, data=spruce.df)}
}
